Fix for game saving

+++ src/Battlescape/Position.h
@@ -92,7 +92,7 @@ public:
 	/// Calculates the distance in 3d.
 	static float distance(Position pos1, Position pos2)
 	{
-		return std::sqrt(distanceSq(pos1, pos2));
+		return static_cast<float>(std::sqrt(distanceSq(pos1, pos2)));
 	}
 
 	/**
